+++ src/taxonomy.h
@@ -7,15 +7,15 @@
 class TaxonomyInput
 {
 private:
-  std::unordered_map<uint64_t, uint64_t> _parent_map;
-  std::unordered_map<uint64_t, uint8_t> _depth_map;
-  std::unordered_map<uint64_t, std::string> _rank_map;
-  std::unordered_map<uint64_t, std::string> _name_map;
+  std::unordered_map<uint32_t, uint32_t> _parent_map;
+  std::unordered_map<uint32_t, uint8_t> _depth_map;
+  std::unordered_map<uint32_t, std::string> _rank_map;
+  std::unordered_map<uint32_t, std::string> _name_map;
 
 public:
   TaxonomyInput(const char *taxonomy_dirpath);
-  uint64_t getParent(uint64_t taxID);
-  std::string getRank(uint64_t taxID);
+  uint32_t getParent(uint32_t taxID);
+  std::string getRank(uint32_t taxID);
   void printTaxonomyInput();
   decltype(_parent_map) &parent_map() { return _parent_map; }
   decltype(_rank_map) &rank_map() { return _rank_map; }
@@ -27,29 +27,29 @@ template<typename T>
 class TaxonomyRecord
 {
 private:
-  uint64_t _num_input;
+  uint32_t _num_input;
   T _num_nodes;
-  uint64_t _full_size;
+  uint32_t _full_size;
   std::unordered_map<T, std::string> _tID_to_rank;
   std::unordered_map<tT, tT> _tID_to_lsroot;
-  std::unordered_map<T, uint64_t> _tID_to_taxID;
-  std::unordered_map<uint64_t, T> _taxID_to_tID;
+  std::unordered_map<T, uint32_t> _tID_to_taxID;
+  std::unordered_map<uint32_t, T> _taxID_to_tID;
   std::unordered_map<std::string, T> _input_to_tID;
   std::unordered_map<T, std::vector<std::string>> _tID_to_input;
   std::vector<T> _parent_vec;
   std::vector<uint8_t> _depth_vec;
   std::unordered_map<T, std::set<T>> _child_map;
-  std::unordered_map<uint64_t, uint64_t> _parent_inmap;
-  std::unordered_map<uint64_t, uint8_t> _depth_inmap;
-  std::unordered_map<uint64_t, std::string> _rank_inmap;
-  std::unordered_map<uint64_t, std::string> _name_inmap;
+  std::unordered_map<uint32_t, uint32_t> _parent_inmap;
+  std::unordered_map<uint32_t, uint8_t> _depth_inmap;
+  std::unordered_map<uint32_t, std::string> _rank_inmap;
+  std::unordered_map<uint32_t, std::string> _name_inmap;
 
 public:
   TaxonomyRecord(const char *input_filepath, TaxonomyInput taxonomy);
   void printTaxonomyRecord();
   T getLowestCommonAncestor(T a, T b);
-  uint64_t taxID_from_tID(T tID);
-  T tID_from_taxID(uint64_t taxID);
+  uint32_t taxID_from_tID(T tID);
+  T tID_from_taxID(uint32_t taxID);
   bool isBasis(T tID);
   decltype(_tID_to_input) &tID_to_input() { return _tID_to_input; }
   decltype(_tID_to_lsroot) &tID_to_lsroot() { return _tID_to_lsroot; }
